% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Used_fn.R
\name{compute_pca}
\alias{compute_pca}
\title{Computing PCA}
\usage{
compute_pca(data_mat, ntop, meta_df)
}
\arguments{
\item{data_mat}{Count matrix rld or vst transformed and normalised}

\item{ntop}{Top n variable features}

\item{meta_df}{Data frame of covariates}
}
\value{
Data frame with PCA loadings and covariates
}
\description{
Computing PCA
}
