% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{count_matrix}
\alias{count_matrix}
\title{Data to be used as examples for the function usage}
\format{
Matrix with count data which is rld transformed having 26730 peaks (rows) and 8 samples (column)
}
\source{
The raw data is present here https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE71596
}
\usage{
count_matrix
}
\description{
Contains epigenetic profiling H3K27ac ChIP-seq count data together with meta file for two tissues PBMC (blood) and synovial fluid (SF) in JIA patients
The data is publicly available
}
\examples{
data(count_matrix)
}
\keyword{datasets}
