% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Used_fn.R
\name{draw_heatmap_top_peaks}
\alias{draw_heatmap_top_peaks}
\title{Clustered heatmap based on variable features}
\usage{
draw_heatmap_top_peaks(data_mat, ntop, col_select, meta_df, rownorm)
}
\arguments{
\item{data_mat}{Count matrix (rld or vst transformed, normalised)}

\item{ntop}{Top n variable features}

\item{col_select}{Vector of annotation column names}

\item{meta_df}{Row annotation data frame}

\item{rownorm}{Z-score across rows. Logical value TRUE/FALSE}
}
\value{
Heatmap based on n variable features
}
\description{
Clustered heatmap based on variable features
}
