% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Used_fn.R
\name{maplot}
\alias{maplot}
\title{MA Plot labelled with significant genes}
\usage{
maplot(res, thresh = 0.05, labelsig = TRUE, textcx = 1, ...)
}
\arguments{
\item{res}{DESeq2 result table of differential analysis}

\item{thresh}{Adjusted p-value threshold. Default is 0.05}

\item{labelsig}{Label genes below threshold. Logical value TRUE or FALSE. Default is TRUE}

\item{textcx}{Text size. Default is 1}

\item{...Other.}{Parameters for base R plot function}
}
\value{
MA plot for differential analysis
}
\description{
MA Plot labelled with significant genes
}
