% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_df}
\alias{meta_df}
\title{Contains metadata associated with epigenetic profiling H3K27ac ChIP-seq count data for two tissues PBMC (blood) and synovial fluid (SF) in JIA patients}
\format{
Data frame with meta data which associated with count data
}
\source{
The raw data is present here https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE71596
}
\usage{
meta_df
}
\description{
Contains metadata associated with epigenetic profiling H3K27ac ChIP-seq count data for two tissues PBMC (blood) and synovial fluid (SF) in JIA patients
}
\examples{
data(meta_df)
}
\keyword{datasets}
