% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Used_fn.R
\name{modify_string}
\alias{modify_string}
\title{Split string based on separator}
\usage{
modify_string(x, delimtter, position, fixed = TRUE)
}
\arguments{
\item{x}{Input string}

\item{delimtter}{Splitting separator}

\item{position}{Position of output string after splitting}

\item{fixed}{Logical value TRUE or FALSE. If TRUE match split exactly otherwise use regular expression. Default is TRUE}
}
\value{
Split string
}
\description{
Split string based on separator
}
