% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Used_fn.R
\name{pca.correlogram}
\alias{pca.correlogram}
\title{Correlogram of pca loadings and covariates}
\usage{
pca.correlogram(
  count_matrix,
  ntop,
  metadata,
  covs,
  scale,
  pcCount,
  logTransform = TRUE
)
}
\arguments{
\item{count_matrix}{Count matrix rld or vst transformed normalised}

\item{ntop}{top n features}

\item{metadata}{Covariates data frame}

\item{covs}{Column names of covariates}

\item{scale}{logical value TRUE or FALSE}

\item{pcCount}{Number of dimension}

\item{logTransform}{Deafault is TRUE}
}
\value{
Heatmap of correlogram between PCA loadings and covariates
}
\description{
Correlogram of pca loadings and covariates
}
