% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{res_df}
\alias{res_df}
\title{Contains differential gene expression data between JIA patients (SF) and healthy controls (PBMC) . It is result table from DESeq2 analysis.}
\format{
Data frame of DESeq2 differential analysis result object
}
\source{
The raw data is present here https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE71595
}
\usage{
res_df
}
\description{
Contains differential gene expression data between JIA patients (SF) and healthy controls (PBMC) . It is result table from DESeq2 analysis.
}
\examples{
data(res_df)
}
\keyword{datasets}
