% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Used_fn.R
\name{to_add_gene_symbol_ENSEMBL}
\alias{to_add_gene_symbol_ENSEMBL}
\title{Mapping ENSEMBL to gene symbols}
\usage{
to_add_gene_symbol_ENSEMBL(input_ensembl_vector)
}
\arguments{
\item{input_ensembl_vector}{Vector of ensembl ids to be mapped}
}
\value{
Data frame with gene symbols, entrezid
}
\description{
Mapping ENSEMBL to gene symbols
}
