% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Used_fn.R
\name{to_do_fisher_exact_test}
\alias{to_do_fisher_exact_test}
\title{Fisher exact test}
\usage{
to_do_fisher_exact_test(
  enrichment_category_groups,
  input_category_groups_to_be_tested,
  x,
  n,
  m,
  k,
  ...
)
}
\arguments{
\item{enrichment_category_groups}{Functional enrichment group name vector (colnames)}

\item{input_category_groups_to_be_tested}{Input category group name vector (rownames)}

\item{x}{overlap of hit list with category of interest}

\item{n}{total no of genes (universe)}

\item{m}{number of genes in a category of interest}

\item{k}{length of hit list}

\item{...}{other parameters to be passed in fisher.test function}
}
\value{
List of contingency table and result of fisher test
}
\description{
Fisher exact test
}
