% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Used_fn.R
\name{to_find_nearest_dist_grange_object}
\alias{to_find_nearest_dist_grange_object}
\title{Nearest distance grane objects with the distance value}
\usage{
to_find_nearest_dist_grange_object(gr1, gr2, gr1_col, gr2_col)
}
\arguments{
\item{gr1}{Genomic Range Object1}

\item{gr2}{Genomic Range Object2}

\item{gr1_col}{Column name for gr1}

\item{gr2_col}{Column name for gr2}
}
\value{
Data frame with genomic range objects with nearest distance
}
\description{
Nearest distance grane objects with the distance value
}
