% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Used_fn.R
\name{to_get_GO_kegg_enrichment}
\alias{to_get_GO_kegg_enrichment}
\title{Title}
\usage{
to_get_GO_kegg_enrichment(
  expressedTarget,
  expressedGenes,
  pval_thresh = 0.05,
  ont = "BP"
)
}
\arguments{
\item{expressedTarget}{Forground genes (ENTREZ IDs)}

\item{expressedGenes}{Background/universe genes (ENTREZ IDs)}

\item{pval_thresh}{Significant p-val threshold. Default is 0.05}

\item{ont}{Default is biological process (BP). Value can be any/all GO term "BP", "MF" or "CC}
}
\value{
List of GO and kegg enrichment
}
\description{
Title
}
