% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Used_fn.R
\name{to_get_overlap_bw_gr_from_peakids}
\alias{to_get_overlap_bw_gr_from_peakids}
\title{Overlap between two sets of genomic coordinates as input}
\usage{
to_get_overlap_bw_gr_from_peakids(
  input_genomic_coord1,
  input_genomic_coord2,
  input1_colname,
  input2_colname
)
}
\arguments{
\item{input_genomic_coord1}{Vector1 containing genomic coordinates}

\item{input_genomic_coord2}{Vector2 containing genomic coordinates}

\item{input1_colname}{Column name of input vector1}

\item{input2_colname}{Column name of input vector2}
}
\value{
Data frame with overlap and length of intersection
}
\description{
Overlap between two sets of genomic coordinates as input
}
