% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Used_fn.R
\name{unique_grobject_metacol}
\alias{unique_grobject_metacol}
\title{Create unique entry gr object with metacolumns}
\usage{
unique_grobject_metacol(genomic_coord, metacolumn_name)
}
\arguments{
\item{genomic_coord}{Vector of genomic coordinates}

\item{metacolumn_name}{Genomic coordinate's column name}
}
\value{
Gr object with unique entries
}
\description{
Create unique entry gr object with metacolumns
}
\examples{
unique_grobject_metacol(genomic_coord = c("chr2:15000-15500","chr5:10001-10122"), metacolumn_name = "Peak_id")
unique_grobject_metacol(genomic_coord = c("chr2:15000-15500","chr5:10001-10122","chr2:15000-15500"), metacolumn_name = "Peak_id")
}
